refactor: rename HttpError to httpError and refactor imports
