+++ src/services/auth.service.js
@@ -1,7 +1,7 @@
 const db = require('../models');
 const bcrypt = require('bcrypt');
 const LoginError = require('../utils/loginError');
-const HttpError = require('../utils/httpError');
+const { HttpError } = require('../utils/httpError');
 const jwt = require('jsonwebtoken');
 const validateUserAndReturnToken = async data => {
   const { email, password } = data;
